refactor: handleDeleteTask moved to TaskList.jsx fix: style button in AddTask.jsx
