+++ src/components/AddTask/AddTask.module.css
@@ -2,7 +2,6 @@
     display: flex;
     justify-content: space-evenly;
     align-items: end;
-
 }
 
 .input {
