+++ src/components/TaskList/TaskList.jsx
@@ -1,6 +1,20 @@
 import TaskItem from "../TasksList/TaskItem.jsx";
+import {deleteTask} from "../../http.js";
 
-export default function TaskList({ isFetching, tasks, handleChangeDataTask, handleDeleteTask }) {
+export default function TaskList({ setIsFetching, setError, fetchTasksByCategories, filter, isFetching, tasks, handleChangeDataTask }) {
+
+    async function handleDeleteTask(id) {
+        setIsFetching(true);
+        try {
+            await deleteTask(id);
+        } catch (error) {
+            setError(error);
+        } finally {
+            setIsFetching(false);
+        }
+        await fetchTasksByCategories(filter);
+    }
+
     return (
         !isFetching && tasks.map(task => <TaskItem
             key={task.id}
